requirements.txt
pandas
biopython
seaborn
matplotlib
tqdm
setuptools
gensim
openpyxl
